refactor: move AddToLogContext example to loggers sub-package
