+++ example_test.go
@@ -4,7 +4,6 @@ import (
 	"context"
 
 	"github.com/go-coldbrew/log"
-	"github.com/go-coldbrew/log/loggers"
 )
 
 func ExampleInfo() {
@@ -27,17 +26,4 @@ func ExampleAddToContext() {
 func ExampleError() {
 	ctx := context.Background()
 	log.Error(ctx, "msg", "database connection failed", "host", "db.internal", "port", 5432, "retry_in", "5s")
-}
-
-// AddToLogContext from the loggers package adds fields at the lower level,
-// useful when building custom logger integrations or interceptors.
-func ExampleAddToLogContext() {
-	ctx := context.Background()
-
-	// Add structured fields that propagate through the interceptor chain
-	ctx = loggers.AddToLogContext(ctx, "service", "payment-gateway")
-	ctx = loggers.AddToLogContext(ctx, "trace_id", "abc-def-123")
-
-	// These fields appear in every log line that uses this context
-	log.Info(ctx, "msg", "charge initiated", "amount", 1999, "currency", "USD")
 }
